Implement Global Player Controls and Update Player Functionality
- Added GlobalPlayerControls component for a mini player interface, featuring track info display, play/pause functionality, and progress bar with seek capability.
- Integrated GlobalPlayerControls into the main app layout for consistent visibility across screens.
- Updated PlayerViewModel to auto-select available players and enhance state management.
- Modified audio model duration types from Int to Float for improved precision.
- Enhanced NowPlayingScreen with detailed player controls and UI improvements.
- Updated MediaType enum to include flow_stream for better media handling.

+++ composeApp/src/commonMain/kotlin/com/mass/client/core/model/Audio.kt
@@ -22,7 +22,7 @@ data class StreamDetails(
     val audio_format: AudioFormat,
     val media_type: MediaType, // Enum
     val stream_title: String? = null,
-    val duration: Int? = null, // seconds
+    val duration: Float? = null, // Changed from Int to Float for decimal durations
     val queue_id: String? = null,
     // Skipping loudness and DSP details for initial model simplicity, can be added later
     // val loudness: Float? = null,
